app/requirements.txt
flask
flask_login
flask_sqlalchemy 
flask_wtf 
flask_bcrypt 
flask_limiter 
flask_mailman
flask_talisman
pyotp 
cryptography
gunicorn
qrcode
zxcvbn
pycryptodome
nh3
markdown
email_validator
pillow
python-dotenv